fix components directory
