+++ src/main/java/models/views/RoundedButton.java
@@ -1,48 +1,125 @@
-package utils;
+package models.views;
 
-import java.awt.BasicStroke;
-import java.awt.Color;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-import java.awt.RenderingHints;
-import java.awt.Shape;
+/**
+ *
+ * @author Markus
+ */
+
+import java.awt.*;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
 import java.awt.geom.Area;
 import java.awt.geom.Rectangle2D;
 import java.awt.geom.RoundRectangle2D;
-import javax.swing.BorderFactory;
-import javax.swing.JTextField;
+import javax.swing.JButton;
 
-public class RoundedTextField extends JTextField{
-    private Color fillColor;
-    private Color lineColor;
+public class RoundedButton extends JButton {
+
+    private boolean over;
+    private Color fill;
+    private Color line;
+
+    private Color fillOriginal;
+    private Color fillOver;
+    private Color fillClick;
+    private Color lineOriginal;
+    private Color lineOver;
     private int strokeWidth;
     private int roundTopLeft = 0;
     private int roundTopRight = 0;
     private int roundBottomLeft = 0;
     private int roundBottomRight = 0;
     
-    public RoundedTextField() {
-        fillColor = new Color(236, 240, 241);
-        lineColor = new Color(52, 152, 219);
+
+    public RoundedButton() {
+        fillOriginal = new Color(52, 152, 219);
+        fillOver = new Color(41, 128, 185);
+        fillClick = new Color(211, 84, 0);
+        lineOriginal = new Color(236, 240, 241);
+        lineOver = new Color(189, 195, 199);
         strokeWidth = 2;
+        fill = fillOriginal;
+        line = lineOriginal;
         setOpaque(false);
-        setBorder(BorderFactory.createEmptyBorder(2, 10, 2, 10));
+        setBorder(null);
+        setFocusPainted(false);
+        setContentAreaFilled(false);
+        setBackground(fillOriginal);
+        setForeground(Color.white);
+        //tambahkan mouse event
+        addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseExited(MouseEvent e) {
+                fill = fillOriginal;
+                line = lineOriginal;
+                over = false;
+            }
+
+            @Override
+            public void mouseEntered(MouseEvent e) {
+                fill = fillOver;
+                line = lineOver;
+                over = true;
+                setCursor(new Cursor(Cursor.HAND_CURSOR));
+            }
+
+            @Override
+            public void mouseReleased(MouseEvent e) {
+                if (over) {
+                    fill = fillOver;
+                    line = lineOver;
+                } else {
+                    fill = fillOriginal;
+                    line = lineOriginal;
+                }
+            }
+
+            @Override
+            public void mousePressed(MouseEvent e) {
+                fill = fillClick;
+            }
+
+        });
     }
 
-    public Color getFillColor() {
-        return fillColor;
+    public Color getFillOriginal() {
+        return fillOriginal;
     }
 
-    public void setFillColor(Color fillColor) {
-        this.fillColor = fillColor;
+    public void setFillOriginal(Color fillOriginal) {
+        this.fillOriginal = fillOriginal;
     }
 
-    public Color getLineColor() {
-        return lineColor;
+    public Color getFillOver() {
+        return fillOver;
     }
 
-    public void setLineColor(Color lineColor) {
-        this.lineColor = lineColor;
+    public void setFillOver(Color fillOver) {
+        this.fillOver = fillOver;
+    }
+
+    public Color getFillClick() {
+        return fillClick;
+    }
+
+    public void setFillClick(Color fillClick) {
+        this.fillClick = fillClick;
+    }
+
+    public Color getLineOriginal() {
+        return lineOriginal;
+    }
+
+    public void setLineOriginal(Color lineOriginal) {
+        this.lineOriginal = lineOriginal;
+    }
+
+    public Color getLineOver() {
+        return lineOver;
+    }
+
+    public void setLineOver(Color lineOver) {
+        this.lineOver = lineOver;
     }
 
     public int getStrokeWidth() {
@@ -52,7 +129,8 @@ public class RoundedTextField extends JTextField{
     public void setStrokeWidth(int strokeWidth) {
         this.strokeWidth = strokeWidth;
     }
-        public int getRoundTopLeft() {
+
+    public int getRoundTopLeft() {
         return roundTopLeft;
     }
 
@@ -151,4 +229,5 @@ public class RoundedTextField extends JTextField{
         area.add(new Area(new Rectangle2D.Double(0, 0, width, height - roundY / 2)));
         return area;
     }
+
 }
