+++ src/main/java/models/views/RoundedPanel.java
@@ -1,135 +1,28 @@
-package utils;
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package models.views;
 
 /**
  *
  * @author Markus
  */
-
-import java.awt.*;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
+import java.awt.Graphics;
+import java.awt.Graphics2D;
+import java.awt.RenderingHints;
+import java.awt.Shape;
 import java.awt.geom.Area;
 import java.awt.geom.Rectangle2D;
 import java.awt.geom.RoundRectangle2D;
-import javax.swing.JButton;
+import javax.swing.JPanel;
 
-public class RoundedButton extends JButton {
+public class RoundedPanel extends JPanel {
 
-    private boolean over;
-    private Color fill;
-    private Color line;
-
-    private Color fillOriginal;
-    private Color fillOver;
-    private Color fillClick;
-    private Color lineOriginal;
-    private Color lineOver;
-    private int strokeWidth;
-    private int roundTopLeft = 0;
-    private int roundTopRight = 0;
-    private int roundBottomLeft = 0;
-    private int roundBottomRight = 0;
-    
-
-    public RoundedButton() {
-        fillOriginal = new Color(52, 152, 219);
-        fillOver = new Color(41, 128, 185);
-        fillClick = new Color(211, 84, 0);
-        lineOriginal = new Color(236, 240, 241);
-        lineOver = new Color(189, 195, 199);
-        strokeWidth = 2;
-        fill = fillOriginal;
-        line = lineOriginal;
+    public RoundedPanel() {
         setOpaque(false);
-        setBorder(null);
-        setFocusPainted(false);
-        setContentAreaFilled(false);
-        setBackground(fillOriginal);
-        setForeground(Color.white);
-        //tambahkan mouse event
-        addMouseListener(new MouseAdapter() {
-            @Override
-            public void mouseExited(MouseEvent e) {
-                fill = fillOriginal;
-                line = lineOriginal;
-                over = false;
-            }
-
-            @Override
-            public void mouseEntered(MouseEvent e) {
-                fill = fillOver;
-                line = lineOver;
-                over = true;
-                setCursor(new Cursor(Cursor.HAND_CURSOR));
-            }
-
-            @Override
-            public void mouseReleased(MouseEvent e) {
-                if (over) {
-                    fill = fillOver;
-                    line = lineOver;
-                } else {
-                    fill = fillOriginal;
-                    line = lineOriginal;
-                }
-            }
-
-            @Override
-            public void mousePressed(MouseEvent e) {
-                fill = fillClick;
-            }
-
-        });
     }
-
-    public Color getFillOriginal() {
-        return fillOriginal;
-    }
-
-    public void setFillOriginal(Color fillOriginal) {
-        this.fillOriginal = fillOriginal;
-    }
-
-    public Color getFillOver() {
-        return fillOver;
-    }
-
-    public void setFillOver(Color fillOver) {
-        this.fillOver = fillOver;
-    }
-
-    public Color getFillClick() {
-        return fillClick;
-    }
-
-    public void setFillClick(Color fillClick) {
-        this.fillClick = fillClick;
-    }
-
-    public Color getLineOriginal() {
-        return lineOriginal;
-    }
-
-    public void setLineOriginal(Color lineOriginal) {
-        this.lineOriginal = lineOriginal;
-    }
-
-    public Color getLineOver() {
-        return lineOver;
-    }
-
-    public void setLineOver(Color lineOver) {
-        this.lineOver = lineOver;
-    }
-
-    public int getStrokeWidth() {
-        return strokeWidth;
-    }
-
-    public void setStrokeWidth(int strokeWidth) {
-        this.strokeWidth = strokeWidth;
-    }
-
+    
     public int getRoundTopLeft() {
         return roundTopLeft;
     }
@@ -165,6 +58,12 @@ public class RoundedButton extends JButton {
         this.roundBottomRight = roundBottomRight;
         repaint();
     }
+
+    private int roundTopLeft = 0;
+    private int roundTopRight = 0;
+    private int roundBottomLeft = 0;
+    private int roundBottomRight = 0;
+
 
     @Override
     protected void paintComponent(Graphics grphcs) {
@@ -229,5 +128,4 @@ public class RoundedButton extends JButton {
         area.add(new Area(new Rectangle2D.Double(0, 0, width, height - roundY / 2)));
         return area;
     }
-
 }
