+++ src/main/java/models/views/RoundedPasswordField.java
@@ -2,12 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package utils;
+package models.views;
 
-/**
- *
- * @author Markus
- */
+import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.RenderingHints;
@@ -15,15 +12,67 @@ import java.awt.Shape;
 import java.awt.geom.Area;
 import java.awt.geom.Rectangle2D;
 import java.awt.geom.RoundRectangle2D;
-import javax.swing.JPanel;
+import javax.swing.BorderFactory;
+import javax.swing.JPasswordField;
+import javax.swing.JTextField;
 
-public class RoundedPanel extends JPanel {
+/**
+ *
+ * @author Markus
+ */
+import java.awt.BasicStroke;
+import java.awt.Color;
+import java.awt.Graphics;
+import java.awt.Graphics2D;
+import java.awt.RenderingHints;
+import java.awt.Shape;
+import java.awt.geom.Area;
+import java.awt.geom.Rectangle2D;
+import java.awt.geom.RoundRectangle2D;
+import javax.swing.BorderFactory;
+import javax.swing.JTextField;
 
-    public RoundedPanel() {
-        setOpaque(false);
-    }
+public class RoundedPasswordField extends JPasswordField{
+    private Color fillColor;
+    private Color lineColor;
+    private int strokeWidth;
+    private int roundTopLeft = 0;
+    private int roundTopRight = 0;
+    private int roundBottomLeft = 0;
+    private int roundBottomRight = 0;
     
-    public int getRoundTopLeft() {
+    public RoundedPasswordField() {
+        fillColor = new Color(236, 240, 241);
+        lineColor = new Color(52, 152, 219);
+        strokeWidth = 2;
+        setOpaque(false);
+        setBorder(BorderFactory.createEmptyBorder(2, 10, 2, 10));
+    }
+
+    public Color getFillColor() {
+        return fillColor;
+    }
+
+    public void setFillColor(Color fillColor) {
+        this.fillColor = fillColor;
+    }
+
+    public Color getLineColor() {
+        return lineColor;
+    }
+
+    public void setLineColor(Color lineColor) {
+        this.lineColor = lineColor;
+    }
+
+    public int getStrokeWidth() {
+        return strokeWidth;
+    }
+
+    public void setStrokeWidth(int strokeWidth) {
+        this.strokeWidth = strokeWidth;
+    }
+        public int getRoundTopLeft() {
         return roundTopLeft;
     }
 
@@ -58,12 +107,6 @@ public class RoundedPanel extends JPanel {
         this.roundBottomRight = roundBottomRight;
         repaint();
     }
-
-    private int roundTopLeft = 0;
-    private int roundTopRight = 0;
-    private int roundBottomLeft = 0;
-    private int roundBottomRight = 0;
-
 
     @Override
     protected void paintComponent(Graphics grphcs) {
